reorganize, command list
