+++ packages/beachball/src/options/getCliOptions.ts
@@ -1,8 +1,9 @@
-import type { Command, OptionValues, OutputConfiguration } from 'commander';
+import { Argument, type Command, type OptionValues, type OutputConfiguration } from 'commander';
 import { findProjectRoot } from 'workspace-tools';
 import { env } from '../env';
 import type { CliOptions, ParsedOptions } from '../types/BeachballOptions';
-import { BeachballCommand, resolveBranchOption, type OptionDefinition } from './cliOptionsHelpers';
+import { BeachballCommand, resolveBranchOption } from './cliOptionsHelpers';
+import { optionDefinitions } from './cliOptionDefinitions';
 
 export interface ProgramContext {
   /** Complete argv (node and script path aren't used but elements must be present) */
@@ -45,80 +46,6 @@ export interface ProgramContext {
 
 /** Command run when none is specified on the command line. */
 const defaultCommand = 'change';
-
-/**
- * Single source of truth for every parseable CLI option: its type, description, short flag, and
- * optional long-flag alias. TypeScript enforces (via the `Record<Exclude<...>>` type) that every
- * `CliOptions` key except the ones filled in elsewhere has an entry here.
- */
-const optionDefinitions: Record<
-  Exclude<keyof CliOptions, 'path' | 'command' | '_extraPositionalArgs' | 'version' | 'help'>,
-  OptionDefinition
-> = {
-  // array options
-  disallowedChangeTypes: { type: 'array', desc: 'change types that are not allowed' },
-  package: {
-    type: 'array',
-    short: 'p',
-    desc: 'force creating a change file for this package (can be specified multiple times)',
-  },
-  scope: {
-    type: 'array',
-    desc: 'only consider package paths matching this pattern (can be specified multiple times; supports negations)',
-  },
-  // boolean options
-  all: { type: 'boolean', desc: 'generate change files for all packages' },
-  bump: { type: 'boolean', desc: 'bump versions during publish' },
-  bumpDeps: { type: 'boolean', desc: 'bump dependent packages during publish' },
-  commit: { type: 'boolean', desc: 'commit change files after "change"' },
-  disallowDeletedChangeFiles: {
-    type: 'boolean',
-    desc: 'verify that no change files were deleted between head and target branch',
-  },
-  fetch: { type: 'boolean', desc: 'fetch from the remote before determining changes' },
-  forceVersions: {
-    type: 'boolean',
-    alias: 'force',
-    desc: "for 'sync': use the version from the registry even if it's older than local",
-  },
-  gitTags: { type: 'boolean', desc: 'create git tags for each published package version' },
-  keepChangeFiles: { type: 'boolean', desc: "don't delete the change files from disk after bumping" },
-  publish: { type: 'boolean', desc: 'publish to the npm registry' },
-  push: { type: 'boolean', desc: 'push changes back to the remote git branch' },
-  verbose: { type: 'boolean', desc: 'print additional information to the console' },
-  yes: { type: 'boolean', short: 'y', desc: 'skip the confirmation prompts' },
-  // number options
-  concurrency: { type: 'number', desc: 'maximum concurrency for write operations such as publishing' },
-  depth: { type: 'number', desc: 'for shallow clones: depth of git history to consider when fetching' },
-  npmReadConcurrency: {
-    type: 'number',
-    desc: 'maximum concurrency for reading package versions from the registry',
-  },
-  gitTimeout: { type: 'number', desc: 'timeout in ms for git push operations' },
-  retries: { type: 'number', desc: 'number of retries for an npm publish before failing' },
-  timeout: { type: 'number', desc: 'timeout in ms for npm operations (other than install)' },
-  // string options
-  access: { desc: 'npm publish access level: "public" or "restricted"' },
-  authType: { short: 'a', desc: 'npm auth type for NPM_TOKEN: "authtoken" or "password"' },
-  branch: { short: 'b', desc: 'target branch from remote (default: git config init.defaultBranch)', omitDefault: true },
-  canaryName: { desc: 'dist-tag and version name to use for canary publishes' },
-  changehint: { desc: 'customized hint message shown when a change file is needed but missing' },
-  changeDir: { desc: 'name of the directory to store change files' },
-  configPath: {
-    short: 'c',
-    alias: 'config',
-    desc: 'custom beachball config path (default: cosmiconfig standard paths)',
-  },
-  dependentChangeType: { desc: 'change type to use for dependent packages (default: patch)' },
-  fromRef: { alias: 'since', desc: 'consider changes or change files since this git ref (branch name, commit SHA)' },
-  message: { short: 'm', desc: 'for "change", the change description; for "publish", the commit message' },
-  packToPath: { desc: 'pack packages to tgz files under this path instead of publishing to npm' },
-  prereleasePrefix: { desc: 'prerelease prefix for packages that will receive a prerelease bump' },
-  registry: { short: 'r', desc: 'npm registry' },
-  tag: { short: 't', desc: 'npm dist-tag for publishes (default: "latest")', omitDefault: true },
-  token: { short: 'n', desc: 'npm auth token (defaults to the NPM_TOKEN environment variable)' },
-  type: { desc: 'type of change: e.g. major, minor, patch, none (instead of prompting)' },
-};
 
 /** Result captured from parsing. */
 interface ParseResult {
@@ -165,7 +92,18 @@ function buildProgram(params: Pick<ProgramContext, 'outputOptions' | 'version'>)
   version && program.version(version);
 
   // The single positional is the command name (any value; validated by the caller/cli.ts).
-  program.argument('[command]', 'beachball command to run');
+  program.addArgument(
+    new Argument('[command]', 'beachball command to run').choices([
+      'change',
+      'check',
+      'bump',
+      'publish',
+      'sync',
+      'config',
+      'canary',
+      'migrate',
+    ])
+  );
 
   let result: ParseResult = { command: defaultCommand, options: {}, extraArgs: [] };
